Add error recovery
Adds a new expression `error!("message" e)` which reports an error and
attempts to recover it with `e`.

Maintains and returns a list of errors which have been recovered.

This enables the parser to report multiple errors, which is useful
in contexts like IDEs and compilers where the user may find it easier
to be presented with all errors rather than just the first.

Changes the public return type of parsers. This is a breaking change,
but clients can easily adapt by invoking `.into_result()` to obtain
the original return type.